Migrated from NPM to Yarn
This commit changes the aluminum repository's supported package manager from NPM to Yarn.

REASONING
==========
Because it allows for synchronous downloads, Yarn will increase the speed at which aluminum dependencies are installed.

COMPATIBILITY
==========
Yarn is compatible with both Travis CI and Dependabot. Thus, the migration should require no (or minimal) further changes to the repository.

CHANGES
==========
This commit replaces NPM's package-lock.json lockfile to Yarn's yarn.lock lockfile. It also changes the "prerequisites" section of the README file to state that Yarn is required.