Added auto-redirection to ./index.html
When a request is made with a URL that ends in /, Wire will automatically serve the index.html file in that directory (if it exists; otherwise, it will respond with a 404 error). An option to disable this feature will be added in the future.
This commit also adds the `$adjrequrl$` variable for use in custom error pages, which is replaced with the request URL with "index.html" appended if applicable (that is, if the original request has a URL ending in /). The default 404 error page has been updated to use this variable rather than `$requrl$`.