Upgraded to Yarn Berry
Aluminum now uses Yarn 2.x (Berry). This commit adds files necessary for Yarn to run, including a .yarnrc.yml file and the code for Yarn itself. The lockfile has also been updated.

This commit also adds more directories and files to the .gitignore file, as recommended by https://yarnpkg.com/getting-started/qa#which-files-should-be-gitignored.

Changes to ensure security and smooth operation of Aluminum with Yarn Berry will likely be committed in the near future.